build success but not runnable add 7zip file from p7zip and 7zip16.04 add third party library and build scripts mv org.tvp.kikikiri2 to org.tvp.kikikiri2n comment jxr, bpg format and oboe lib change some private class fix import of KR2Activaty fix some building problem

+++ src/core/base/7zArchive.cpp
@@ -10,8 +10,8 @@ extern "C" {
 #include "StorageImpl.h"
 
 static ISzAlloc allocImp = {
-	[](void *p, size_t size) -> void *{ return malloc(size); },
-	[](void *p, void *addr) { free(addr); }
+	[](void* p, unsigned long size) -> void *{ return malloc(size); },
+	[](void* p, void *addr) { free(addr); }
 };
 
 class SevenZipStreamWrap {
